Move ClientRequestAdapter and ClientResponseAdapter.
Move from brave-interfaces module to brave-http-client module.
These are specific to http client integration so move them to
appropriate sub module.
