+++ brave-http-client/src/main/java/com/github/kristofa/brave/client/ClientResponseAdapter.java
@@ -1,4 +1,4 @@
-package com.github.kristofa.brave;
+package com.github.kristofa.brave.client;
 
 public interface ClientResponseAdapter {
     int getStatusCode();
