+++ brave-http-client/src/main/java/com/github/kristofa/brave/client/ClientResponseInterceptor.java
@@ -2,7 +2,6 @@ package com.github.kristofa.brave.client;
 
 import org.apache.commons.lang.Validate;
 
-import com.github.kristofa.brave.ClientResponseAdapter;
 import com.github.kristofa.brave.ClientTracer;
 
 /**
