+++ brave-http-client/src/test/java/com/github/kristofa/brave/client/ClientRequestHeadersTest.java
@@ -11,7 +11,6 @@ import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
 import com.github.kristofa.brave.BraveHttpHeaders;
-import com.github.kristofa.brave.ClientRequestAdapter;
 import com.github.kristofa.brave.ClientTracer;
 import com.github.kristofa.brave.SpanId;
