+++ brave-jaxrs2/src/main/java/com/github/kristofa/brave/jaxrs2/JaxRS2ClientResponseAdapter.java
@@ -1,6 +1,6 @@
 package com.github.kristofa.brave.jaxrs2;
 
-import com.github.kristofa.brave.ClientResponseAdapter;
+import com.github.kristofa.brave.client.ClientResponseAdapter;
 
 import javax.ws.rs.client.ClientResponseContext;
