+++ brave-jersey/src/main/java/com/github/kristofa/brave/jersey/JerseyClientResponseAdapter.java
@@ -1,6 +1,6 @@
 package com.github.kristofa.brave.jersey;
 
-import com.github.kristofa.brave.ClientResponseAdapter;
+import com.github.kristofa.brave.client.ClientResponseAdapter;
 import com.sun.jersey.api.client.ClientResponse;
 
 class JerseyClientResponseAdapter implements ClientResponseAdapter {
